+++ index.html
@@ -227,7 +227,6 @@
     </section>
   </template>
 
-  <script src="js/main.js"></script>
-  <script src="js/functions.js"></script>
+  <script src="./js/main.js" type="module"></script>
 </body>
 </html>
