+++ js/render.js
@@ -1,0 +1,5 @@
+const render = () => void 0;
+
+export { render };
+
+// console.log('загружен render');
