+++ js/slider.js
@@ -1,0 +1,5 @@
+const slider = () => void 0;
+
+export { slider };
+
+// console.log('загружен slider');
