+++ js/validation.js
@@ -1,0 +1,5 @@
+const validation = () => void 0;
+
+export { validation };
+
+// console.log('загружен validation');
